// web/packages/teleport/src/Integrations/Enroll/Cloud/RegionMultiSelector/types.ts
/**
 * Teleport
 * Copyright (C) 2025  Gravitational, Inc.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

import { Regions as AwsRegion } from 'teleport/services/integrations';

export type RegionId = AwsRegion; // | AzureRegion | GcpRegion, etc

export interface Region {
  id: RegionId;
  name: string;
}

export interface RegionGroup {
  name: string;
  regions: readonly Region[];
}